requirements.txt
asgiref==3.11.0
brotli==1.2.0
click==8.3.1
Django==6.0
django-environ==0.12.0
django-guardian==3.2.0
gunicorn==23.0.0
h11==0.16.0
packaging==25.0
Pygments==2.19.2
PyJWT==2.10.1
sqlparse==0.5.4
uvicorn==0.38.0
whitenoise==6.11.0
